ENH: Added command line package for dti processing.
This package requires the boost library and support is added in CMake.  The main two executables in this package are dtiestim and dtiprocess.  dtiestim is a command line tool for estimating a tensor field from a set of diffusion weighted images.  dtiprocess is a command line tool for processing tensor fields.  dtiprocess supports creating scalar, color images from the tensor field and transforming tensor fields.