BUG: Addition of the macro unsetForSlicer
If Slicer is not configured with the system compiler and default compilation flags, the extensions will be configured with the default compiler and the default flags, which will create errors. To avoid that, we remove the compiler and the flags before including Slicer which sets the flags and the compiler to the values that were used by Slicer.