ENH: Addition of an option to shift negative eigenvalues to positive values.
This option was available in DWIToDTIEstimation, a 3DSlicer module. All the eigenvalues are shifted by the opposite of the smallest (most negative) eigenvalue such that all the eigenvalues of the tensor are positive. Additionally, we let the user set a coefficient that will multiply the eigenvalue used for the shift. This allows to give an "epsilon" to enforce strict positivity of the eigenvalues without giving a strict value to the "epsilon".