ENH: When building DTIProcess as a Slicer extension, do not specify the inner-build directory as the "packaging" directory.
When setting the CMake arguments for Slicer to a package extension, the scripts adds many rules to build the extension. Some of the rules reset all the CMake variables. Because of this, if an extension uses the Superbuild mechanism to specify different libraries from the Slicer libraries, these new library paths will be lost upon reinitialization. To avoid this problem, we specify the outer-build directory as the directory that Slicer should use to package the extension. This require to install in a known directory the executable we want to package as well as the tests.