ENH: Do not rebuild ITK, VTK, or SlicerExecutionModel
When compiling niralUtilities, it is using the Superbuild system and therefore makes sure that all libraries are available, otherwise it rebuilds them. ITK, VTK, and SlicerExecutionModel are already included in DTIProcess. Therefore there is no need to recompile them. To avoid recompiling them, we set the corresponding USE_SYSTEM_ variables to ON that the superbuild system recognizes.