BUG: Fix loading of CLI modules when built as an extension.
This commit partially fixes issue #14. It only addresses the loading
of CLI executables built in the Applications folder. The CLI
build in the "niral_utilities" project [1] would have to be updated
to either

 (1) use the macro "SEMMacroBuildCLI" so that configuring the
"niral_utilities" project with SlicerExecutionModel_DEFAULT_CLI_(EXECUTABLE|LIBRARY)_LINK_FLAGS
is effective

or

 (2) explicitly set the linker flags on the executable.

Approach (1) is definitively recommended.